Elimiated the reason for the 'cannot add event because data type is unknown' warning.
On startup, it's unclear who connects first.
In case a channel was disabled before it actually connected,
we didn't add the 'disabled'.
Now, it's added, assuming 'double' type.

Later, on connection, when allocating the 'correct' circular buffer,
all 'info' values that might have accumulated
are copied into the new buffer.